requirements.txt
# Base LangChain functionality
langchain>=0.1.0
langchain-core>=0.1.7
langchain-community>=0.0.10
langchain-experimental>=0.0.47
langchain-ollama>=0.0.10

# Graph-based workflows
langgraph>=0.0.12
graphviz>=0.20.1

# Database clients
opensearch-py>=2.0.0

# Environment and configuration
python-dotenv>=1.0.0
pydantic>=2.6.4

# HTTP clients
requests>=2.31.0
httpx>=0.26.0
aiohttp>=3.9.3
urllib3>=1.26.0

# Data processing
pandas>=2.0.0
numpy>=1.26.0
pyarrow>=19.0.0

# Date handling
python-dateutil>=2.8.2
pytz>=2024.1

# Utils
typing-extensions>=4.9.0
rich>=13.7.0
tenacity>=8.2.3
importlib-metadata>=6.0.0